requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.4.1
crispy-bootstrap5==2024.2
distlib==0.3.8
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1
django-allauth==65.3.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-email==0.1.10
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.15.4
idna==3.10
Markdown==3.7
MouseInfo==0.1.3
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-decouple==3.8
pytweening==1.2.0
requests==2.32.3
setuptools==73.0.1
sqlparse==0.5.1
sslcommerz==0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
virtualenv==20.26.3
